Ordered builder containers + README additions
Sadly, the best option available to us right now to ensure that the
docker container initializes before the celery process is to reach back
into our hardcoded environment variable patching mechanisms and define a corner case for the
builder StatefulSet.

It's sad, but it works, and I'm tired of struggling against kustomize to
do simple things.